refactor: new notifications module and move notifications.js to it
